feat: paginate sandbox listings
Regenerate gateway protos with the ListSandboxes pagination fields and
route Sandbox.list() through a shared paginate_async helper. Reuse the
helper for discovery list RPCs so all paginated list calls share
deadline, loop-detection, and max-page handling.

This prevents high-volume listings from silently stopping at the
server's first-page cap.
